msm: mdss: fix fence timeout for 2 buff cases
In cases where there are 2 buffer apps, there is a
race condition, where mdp would be waiting for gfx
fence. The condition is as follows:
1. FB driver gets a kickoff1, and the internal kernel
thread triggers the data transfer for this buffer1.
2. FB driver gets a second kickoff2 for the buffer2.
3. The data transfer of buffer1 finishes and we receive
a ping pong done interrupt, but since the kickoff2 is
already on-going at this time, the pp-done interrupt
doesn't schedule the work item to release the fence
and instead, the driver will be releasing the fence
from the kickoff2 thread.
4. The issue happens at this time, because the kickoff2
thread is now waiting for the fence from gfx for the
buffer2 before releasing the previous fence from
buffer1, but gfx is waiting for the fence from the
kickoff1.

Mot-CRs-fixed: (CR)
QCOM case/CR: 02814457/2034044

Change-Id: I4523bbd1eaa9ae0e3e47101dc08be61bd84b5750
Signed-off-by: Ingrid Gallardo <ingridg@codeaurora.org>
Reviewed-on: https://gerrit.mot.com/1165997
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Guobin Zhang <zhanggb@motorola.com>
Submit-Approved: Shanshan Dai <daiss1@lenovo.com>
Signed-off-by: laststandrighthere <laststandrighthere@gmail.com>
Signed-off-by: Lau <laststandrighthere@gmail.com>